% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achisqaprox.R
\name{achisqaprox}
\alias{achisqaprox}
\title{Monte Carlo Approximation of the Chi-Square Distribution}
\usage{
achisqaprox(m, v)
}
\arguments{
\item{m}{Number of Monte Carlo replications.}

\item{v}{Degrees of freedom of the Chi-square distribution.}
}
\value{
A numeric vector of length \code{m} containing simulated values
from a Chi-square distribution with \code{v} degrees of freedom.
}
\description{
using standard normal random variables. Each simulated value is obtained
as the sum of squared standard normal variables generated with
\code{normbox}.
}
\details{
If Z₁, Z₂, ..., Zᵥ are independent standard normal random variables,
then the random variable ∑ Zᵢ² follows a Chi-square distribution with
\code{v} degrees of freedom.

This function uses the \code{normbox} function from the
\code{erandom} package to generate the standard normal samples.
}
\examples{
library(erandom)
x <- achisqaprox(m = 5000, v = 5)
hist(x, probability = TRUE, main = "Chi-square approximation")
curve(dchisq(x, df = 5), col = "red", add = TRUE)

}
\seealso{
\code{\link[stats]{dchisq}}, \code{normbox}
}
