% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glewis.R
\name{glewis}
\alias{glewis}
\title{Pseudorandom Number Generator using the Lewis Method}
\usage{
glewis(n)
}
\arguments{
\item{n}{Integer. Number of pseudorandom values to generate.}
}
\value{
A numeric vector of length \code{n} containing pseudorandom numbers
uniformly distributed.
}
\description{
Implements the *Multiplicative Congruential Generator*
for capacity of the IBM System/360 computers create a sequence of
pseudo-random numbers.
}
\details{
The generator produces a sequence of pseudorandom numbers.
\deqn{X_{i+1} = (a \times X_i) \mod m}
where \eqn{a = 7^5} and \eqn{m = 2^{31} - 1 = 2147483647}.

Here, \eqn{m} is a *prime number*, and \eqn{a = 16807}
is a *positive primitive root* modulo \eqn{m}, which guarantees a *maximal period*
of 2147483647.
}
\examples{
# Generate 5 pseudorandom numbers using the Lewis method
glewis(5)

}
\references{
Lewis, P. A. W., Goodman, A. S., & Miller, J. M. (1969).
A pseudo-random number generator for the System/360.
IBM Systems Journal, 8(2), 136–146.
}
