% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Period.R
\name{Period}
\alias{Period}
\title{Period}
\usage{
Period(x)
}
\arguments{
\item{x}{a random sequence}
}
\value{
Returns a numeric value representing the number of unique, non-repeated
elements in the input sequence or vector.In other words, the count of values
that appear exactly once in the sequence.
}
\description{
This function counts the number of unique, non-repeated elements in a sequence or vector.
In other words, it identifies and counts the values that appear exactly once within
the sequence.
}
\examples{
x <- runif(n = 100, min = 0, max = 1)
Period(x)

}
\references{
R Core Team (2024). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing, Vienna, Austria.
}
