% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniftest.R
\name{uniftest}
\alias{uniftest}
\title{Test for Uniformity Unif(0, 1)}
\usage{
uniftest(u)
}
\arguments{
\item{u}{a random sequence}
}
\value{
The function returns the test statistic, the p-value and the degree
of freedom of a chi-square goodness of fit test to prove
uniformity U(0, 1) from a sequence.
}
\description{
This function implements the chi-square goodness of fit test to prove
uniformity U(0, 1) from a sequence
}
\examples{
u <- runif(n = 100, min = 0, max = 1)
uniftest(u)

}
\references{
Andreis, F. (2014). R and Introduction to Simulation.
}
