% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independ.R
\name{independ}
\alias{independ}
\title{Test of Independence for Random Sequences}
\usage{
independ(x, lags)
}
\arguments{
\item{x}{A random sequence}

\item{lags}{The number of lags (autocorrelation terms) to include in the test statistic.}
}
\value{
The function returns the test statistic (Q), the degrees of freedom,
and the p-value of a chi-square test for serial independence based on
sample autocorrelations up to the specified number of lags.
}
\description{
This function performs the Ljung-Box test to verify whether a sequence of random
numbers is truly independent across its lags. It takes the series of generated
values and calculates the autocorrelations between the numbers at different lags
up to a certain limit. Using these autocorrelations, it calculates the Q statistic,
which follows a chi-square distribution, to determine whether there is dependence
between the values of the sequence.
}
\examples{
set.seed(123)
x <- rnorm(100)
independ(x, lags = 10)

}
\references{
Burns, P. (2002). Robustness of the Ljung-Box test and its rank equivalent. Available at SSRN 443560.
}
